+++ sdk/monitor/azure-monitor-query/tests/test_logs_client.py
@@ -18,7 +18,7 @@ from azure.monitor.query import (
     LogsQueryPartialResult,
     LogsQueryStatus,
 )
-from azure.monitor.query._helpers import native_col_type
+from azure.monitor.query._utils import native_col_type
 from azure.monitor.query._version import VERSION
 
 from base_testcase import AzureMonitorQueryLogsTestCase
@@ -196,7 +196,6 @@ class TestLogsClient(AzureMonitorQueryLogsTestCase):
         assert response is not None
         assert len(response.tables[0].rows) == 2
 
-    @pytest.mark.skip("Flaky deserialization issues with msrest. Re-enable after removing msrest dependency.")
     @pytest.mark.live_test_only("Issues recording dynamic 'id' values in requests/responses")
     def test_logs_query_batch_additional_workspaces(self, monitor_info):
         client = self.get_client(LogsQueryClient, self.get_credential(LogsQueryClient))
@@ -298,10 +297,11 @@ class TestLogsClient(AzureMonitorQueryLogsTestCase):
 
     def test_client_different_endpoint(self):
         credential = self.get_credential(LogsQueryClient)
-        endpoint = "https://api.loganalytics.azure.cn/v1"
+        endpoint = "https://api.loganalytics.azure.cn"
         client = LogsQueryClient(credential, endpoint=endpoint)
 
         assert client._endpoint == endpoint
+        assert client._client._config.authentication_policy
         assert "https://api.loganalytics.azure.cn/.default" in client._client._config.authentication_policy._scopes
 
     def test_client_user_agent(self):